Move Orgalog related files
Moves the following into the orgalog dir:
- model OrgaLog
- project info tags for the orgalogs on the project view
- templates for tags, rendering and orgalog view

Minor fix in render_orgalogs; renders link to detail page again
