+++ bp/templates/bp/orgalogs/render_orga_logs.html
@@ -9,7 +9,7 @@
             <td>{{ orgalog.problems }}</td>
             <td>{{ orgalog.text | linebreaksbr | truncatewords:50 }}</td>
             <td>
-                {% if user.is_staff %}
+                {% if can_view_orgalog_detail %}
                     <a href="{% url "bp:orga_log_detail" pk=orgalog.pk %}">{% fa5_icon "link" "fas" %}</a>
                 {% endif %}
             </td>
