--- bp/templates/bp/project_info_misc_orgalog_content.html
@@ -1,5 +1,0 @@
-{% if orga_log_count > 0 %}
-    {% include "bp/render_orga_logs.html" with logs=orga_logs %}
-{% else %}
-
-{% endif %}
