+++ bp/templatetags/tags_project_info_misc.py
@@ -92,22 +92,6 @@ class ProjectInfoTabs():
         heapq.heappush(ProjectInfoTabs.registered_tabs, (priority, tag_names))
 
 
-orgalog_tab = ProjectInfoTabTags(priority=2)
-
-@orgalog_tab.register_description('bp/project_info_misc_orgalog_desc.html')
-def orgalog_description(project):
-    return {
-        'orga_log_count' : project.orgalog_set.all().count(),
-    }
-
-@orgalog_tab.register_content('bp/project_info_misc_orgalog_content.html')
-def orgalog_content(project):
-    return {
-        'orga_logs'      : project.orgalog_set.all(),
-        'orga_log_count' : project.orgalog_set.all().count(),
-    }
-
-
 grading_tab = ProjectInfoTabTags(priority=3)
 
 @grading_tab.register_description('bp/project_info_misc_grading_desc.html')
